BUGFIX: If we don't set bi_vcnt correctly, block layer may merge two bios that are contiguous sector-wise but are not contiguous page-wise.
For example, if one clone of a bio starts at sector 0 and has a size of 4096
bytes, its bio_vec points to page 10, and another clone of the same bio starts
at sector 8 (i.e. it is contiguous with the previous one) and has a size of 4096
bytes, but its bio_vec points to page 15, the block layer will merge these two
and write to pages 10 and 11, instead of 10 and 15 as expected.  To fix that, we
have to set the bi_vcnt field correctly so that block layer does not assume that
pages are physically contiguous as well.