BUGFIX: This patch fixes two bugs.
First bug is calling reinit_completion only in the case of sc->error being
-EINPROGRESS.  This is wrong, since, it is possible that by the time we reach
the check for -EINPROGRESS, requests have completed and sc->error is 0.  In that
case, was not calling reinit_completion which was causing wait_for_completion to
fire immediately in the next iteration of the loop, before the requests of the
next stage completed.

As a fix, we call reinit_completion on every iteration of the state machine --
while it is redundant for some cases, it is a simple fix that works.

The second bug is, in every stage of RMW was launching requests first and
setting the sc->error to -EINPROGRESS later.  It is possible that by the time we
reach the line that sets sc->error to -EINPROGRESS, all of the requests are
complete, and sc->error is 0.  Setting to sc->error to -EINPROGRESS after leads
to missing a completed operation.

As a fix, set sc->error to -EINPROGRESS before launching the requests.